+++ src/Models/FlowResponse.php
@@ -11,7 +11,7 @@ use Illuminate\Database\Eloquent\SoftDeletes;
 use Override;
 use RobinsonRyan\FormFlow\Enums\ActorType;
 use RobinsonRyan\FormFlow\Enums\ResponseStatus;
-use RobinsonRyan\FormFlow\Traits\HasConfigurableUuid;
+use RobinsonRyan\FormFlow\Traits\ConfiguresIdentifiers;
 
 /**
  * @property string $id
@@ -37,7 +37,7 @@ use RobinsonRyan\FormFlow\Traits\HasConfigurableUuid;
  */
 final class FlowResponse extends Model
 {
-    use HasConfigurableUuid;
+    use ConfiguresIdentifiers;
     use SoftDeletes;
 
     protected $guarded = [];